% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dleierchen}
\alias{dleierchen}
\title{Full text of the \emph{D'Léierchen} poem in Luxembourguish}
\format{A data frame with 6471 rows and 3 variables:
\describe{
  \item{\code{line}}{Original text}
  \item{\code{nb_line}}{Line number}
}}
\source{
\url{https://data.public.lu/fr/datasets/the-works-in-luxembourguish-of-michel-rodange/}
}
\usage{
dleierchen
}
\description{
A dataset containing the full text of Michel Rodange's
poem, \emph{D'Léierchen, Dem Léiweckerche säi Lidd} in
Luxembourguish.
}
\keyword{datasets}
