% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renert}
\alias{renert}
\title{Full text of the \emph{Renert} in Luxembourguish}
\format{A data frame with 6471 rows and 3 variables:
\describe{
  \item{\code{line}}{Original text}
  \item{\code{gesank}}{\emph{Song}; the text is subdivided in songs, called \emph{gesank} in Luxembourguish}
  \item{\code{nb_line}}{Line number}
}}
\source{
\url{https://data.public.lu/fr/datasets/the-works-in-luxembourguish-of-michel-rodange/}
}
\usage{
renert
}
\description{
A dataset containing the full text of Michel Rodange's
epic, \emph{Renert oder de Fuuß am Frack an a Ma'nsgrëßt} in
Luxembourguish.
}
\keyword{datasets}
