% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigfrid}
\alias{sigfrid}
\title{Full text of the \emph{Sigfrid} play in Luxembourguish}
\format{A data frame with 6471 rows and 3 variables:
\describe{
  \item{\code{line}}{Original text}
  \item{\code{act}}{Section of the play}
  \item{\code{nb_line}}{Line number}
}}
\source{
\url{https://data.public.lu/fr/datasets/the-works-in-luxembourguish-of-michel-rodange/}
}
\usage{
sigfrid
}
\description{
A dataset containing the full text of Michel Rodange's
play, \emph{Dem Grow Sigfrid seng Goldkuommer} in
Luxembourguish.
}
\keyword{datasets}
